CDLS-02: refactor users
