+++ src/main/java/com/codeless/api/automation/controller/UserController.java
@@ -1,12 +1,9 @@
-package com.codeless.api.automation.controller.security;
+package com.codeless.api.automation.controller;
 
-import com.codeless.api.automation.dto.AppUser;
-import com.codeless.api.automation.dto.View;
-import com.codeless.api.automation.service.security.UserService;
-import com.fasterxml.jackson.annotation.JsonView;
+import com.codeless.api.automation.dto.UserRegistration;
+import com.codeless.api.automation.service.UserService;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,8 +16,7 @@ public class UserController {
   private final UserService userService;
 
   @PostMapping(value = "/registration", produces = MediaType.APPLICATION_JSON_VALUE)
-  @JsonView(View.ALLOWED.class)
-  public AppUser register(@Valid @RequestBody AppUser appUser) {
-    return userService.saveUser(appUser);
+  public void register(@Valid @RequestBody UserRegistration userRegistration) {
+    userService.saveUser(userRegistration);
   }
 }
