+++ src/main/java/com/codeless/api/automation/dto/UserRegistration.java
@@ -1,6 +1,5 @@
 package com.codeless.api.automation.dto;
 
-import com.fasterxml.jackson.annotation.JsonView;
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Size;
@@ -9,8 +8,7 @@ import lombok.experimental.Accessors;
 
 @Data
 @Accessors(chain = true)
-@JsonView(View.ALLOWED.class)
-public class AppUser {
+public class UserRegistration {
 
   @NotEmpty
   @Email
@@ -18,6 +16,5 @@ public class AppUser {
 
   @Size(max = 100, min = 5, message = "Invalid password size. Min - 5, Max - 100.")
   @NotEmpty
-  @JsonView(View.NOT_ALLOWED.class)
   private String password;
 }
