+++ src/main/java/com/codeless/api/automation/service/UserService.java
@@ -1,0 +1,8 @@
+package com.codeless.api.automation.service;
+
+import com.codeless.api.automation.dto.UserRegistration;
+
+public interface UserService {
+
+  void saveUser(UserRegistration userRegistration);
+}
