--- src/main/java/com/codeless/api/automation/service/security/UserService.java
@@ -1,8 +1,0 @@
-package com.codeless.api.automation.service.security;
-
-import com.codeless.api.automation.dto.AppUser;
-
-public interface UserService {
-
-  AppUser saveUser(AppUser appUser);
-}
